Skrivit in lite felhantering i deleteCarById. Lagt till message funktion i felmeddelanden.

+++ src/main/java/com/example/car_spring_boot/ControllerServiceCar/CarService.java
@@ -53,6 +53,12 @@ public class CarService {
 
 
     public Void deleteCarById(Integer id) {
+        Optional<CarEntity> OpCar = carRepository.findById(id);
+
+        if (OpCar.isEmpty()){
+            throw new IllegalStateException("Id finns inte");
+        }
+
         carRepository.deleteById(id);
         return null;
     }
